requirements.txt
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.0.1
pillow==11.0.0
django-cors-headers==4.6.0
drf-yasg==1.21.8
celery==5.4.0
redis==5.2.0
eventlet==0.37.0
django-celery-beat==2.7.0
flake8==7.1.1
requests==2.32.3
coverage==7.6.4